feature: add :feature:makecolor
