+++ app/build.gradle.kts
@@ -79,8 +79,9 @@ dependencies {
     implementation(project(":core:resource"))
     implementation(project(":core:ui"))
     implementation(project(":core:util"))
-    implementation(project(":feature:settings"))
+    implementation(project(":feature:makecolor"))
     implementation(project(":feature:preview"))
+    implementation(project(":feature:settings"))
     compileOnly(libs.firebase.crashlytics.gradle)
     implementation(libs.accompanist.navigationAnimation)
     implementation(libs.accompanist.systemuicontroller)
