+++ feature/makecolor/build.gradle.kts
@@ -1,0 +1,44 @@
+plugins {
+    id(libs.plugins.android.library.get().pluginId)
+    id(libs.plugins.kotlin.android.get().pluginId)
+    id(libs.plugins.hilt.get().pluginId)
+    id(libs.plugins.kotlin.kapt.get().pluginId)
+}
+
+android {
+    compileSdk = 33
+    namespace = "kosenda.makecolor.feature.makecolor"
+    defaultConfig {
+        minSdk = 24
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
+    }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = libs.versions.androidxComposeCompiler.get()
+    }
+}
+
+dependencies {
+    implementation(project(":core:data"))
+    implementation(project(":core:domain"))
+    implementation(project(":core:model"))
+    implementation(project(":core:ui"))
+    implementation(project(":core:util"))
+    implementation(project(":feature:preview"))
+    implementation(libs.accompanist.systemuicontroller)
+    implementation(libs.androidx.appcompat)
+    implementation(libs.androidx.compose.material3)
+    implementation(libs.androidx.navigation.compose)
+    implementation(libs.androidx.compose.ui.tooling)
+    implementation(libs.androidx.lifecycle.viewModelCompose)
+    implementation(libs.androidx.palette)
+    implementation(libs.compose.color.picker)
+    implementation(libs.hilt.android)
+    implementation(libs.timber)
+    kapt(libs.hilt.compiler)
+}
