+++ feature/makecolor/src/main/java/kosenda/makecolor/feature/makecolor/screen/PictureScreen.kt
@@ -1,10 +1,14 @@
-package kosenda.makecolor.view.screen
+package kosenda.makecolor.feature.makecolor.screen
 
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
@@ -16,61 +20,67 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.stringArrayResource
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import kosenda.makecolor.R
 import kosenda.makecolor.core.model.data.ColorData
+import kosenda.makecolor.core.ui.feature.common.ImagePicker
 import kosenda.makecolor.core.ui.feature.common.button.FloatingAddButton
 import kosenda.makecolor.core.ui.feature.common.button.TonalButton
 import kosenda.makecolor.core.ui.feature.common.card.ColorValueTextsCard
 import kosenda.makecolor.core.ui.feature.common.card.DisplayColorCard
-import kosenda.makecolor.core.ui.feature.common.card.RandomColorsCard
-import kosenda.makecolor.core.ui.feature.common.card.SpinnerCard
+import kosenda.makecolor.core.ui.feature.common.card.PaletteCircleCard
+import kosenda.makecolor.core.ui.feature.common.card.SelectImageCard
 import kosenda.makecolor.core.ui.feature.common.topbar.TopBar
 import kosenda.makecolor.core.ui.feature.theme.MakeColorTheme
+import kosenda.makecolor.feature.makecolor.R
+import kosenda.makecolor.feature.makecolor.viewmodel.PictureViewModel
+import kosenda.makecolor.feature.makecolor.viewmodel.PictureViewModelImpl
+import kosenda.makecolor.feature.makecolor.viewmodel.PreviewPictureViewModel
 import kosenda.makecolor.feature.preview.PreviewSurface
-import kosenda.makecolor.view.content.GoogleAd
-import kosenda.makecolor.viewmodel.PreviewRandomViewModel
-import kosenda.makecolor.viewmodel.RandomViewModel
-import kosenda.makecolor.viewmodel.RandomViewModelImpl
 
 @Composable
-fun RandomScreen(
-    viewModel: RandomViewModelImpl,
+fun PictureScreen(
+    viewModel: PictureViewModelImpl,
     onClickMenu: () -> Unit,
     onClickInfo: () -> Unit,
     onClickFloatingButton: (ColorData) -> Unit,
     onClickDisplayColor: (ColorData) -> Unit,
+    googleAd: @Composable () -> Unit = {},
 ) {
-    RandomScreenContent(
+    PictureScreenContent(
         viewModel = viewModel,
         onClickMenu = onClickMenu,
         onClickInfo = onClickInfo,
         onClickFloatingButton = onClickFloatingButton,
         onClickDisplayColor = onClickDisplayColor,
+        googleAd = googleAd,
     )
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun RandomScreenContent(
-    viewModel: RandomViewModel,
+fun PictureScreenContent(
+    viewModel: PictureViewModel,
     onClickMenu: () -> Unit,
     onClickInfo: () -> Unit,
     onClickFloatingButton: (ColorData) -> Unit,
     onClickDisplayColor: (ColorData) -> Unit,
-    googleAd: @Composable () -> Unit = { GoogleAd() },
+    googleAd: @Composable () -> Unit = {},
 ) {
     val uiState by viewModel.uiState.collectAsState()
-    val scrollState = androidx.compose.foundation.rememberScrollState()
+    val context = LocalContext.current
+    val scrollState = rememberScrollState()
     val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(rememberTopAppBarState())
     val horizontalPadding = 16.dp
-    val randomTypeArray = stringArrayResource(id = R.array.random_category)
-    val screenWidth = LocalConfiguration.current.screenWidthDp.dp - horizontalPadding * 2
-    val size = (screenWidth - horizontalPadding * 2 + 8.dp) / 5 - 8.dp
+    val launcher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.GetContent(),
+        onResult = {
+            it?.let { uri -> viewModel.makeBitmapAndPalette(uri = uri, context = context) }
+        },
+    )
 
     Scaffold(
         modifier = Modifier.fillMaxSize(),
@@ -102,29 +112,36 @@ fun RandomScreenContent(
                 colorData = uiState.colorData,
                 onClickDisplayColor = onClickDisplayColor,
             )
-            Column {
-                SpinnerCard(
-                    modifier = Modifier.padding(top = 8.dp),
-                    selectedText = randomTypeArray[uiState.selectRandomType.index],
-                    categoryName = stringResource(id = R.string.category),
-                    onSelectedChange = viewModel::updateRandomType,
-                    displayItemList = randomTypeArray,
-                )
-                RandomColorsCard(
-                    randomRGBColors = uiState.randomRGBColors,
-                    size = size,
+            if (uiState.bitmap == null) {
+                SelectImageCard(onClick = { launcher.launch("image/*") })
+            } else {
+                PaletteCircleCard(
+                    paletteColors = uiState.paletteColors,
+                    horizontalPadding = horizontalPadding,
                     updateColorData = viewModel::updateColorData,
                 )
-                TonalButton(
-                    modifier = Modifier
-                        .padding(top = 16.dp, bottom = 8.dp)
-                        .fillMaxSize(),
-                    text = stringResource(id = R.string.output_random),
-                    onClick = viewModel::outputRandomColors,
-                )
+                uiState.bitmap?.let { btm ->
+                    ImagePicker(
+                        btm = btm,
+                        horizontalPadding = horizontalPadding,
+                        updateColorData = viewModel::updateColorData,
+                    )
+                    Row(
+                        modifier = Modifier.padding(top = 16.dp),
+                    ) {
+                        TonalButton(
+                            text = stringResource(id = R.string.reset),
+                            painter = painterResource(id = R.drawable.baseline_cached_24),
+                            onClick = viewModel::resetImage,
+                        )
+                        Spacer(modifier = Modifier.weight(1f))
+                    }
+                }
             }
-
-            ColorValueTextsCard(colorData = uiState.colorData)
+            ColorValueTextsCard(
+                modifier = Modifier.padding(top = 8.dp),
+                colorData = uiState.colorData,
+            )
             Spacer(modifier = Modifier.height(150.dp))
         }
     }
@@ -132,11 +149,11 @@ fun RandomScreenContent(
 
 @Preview
 @Composable
-private fun PreviewRandomScreenContent_Light() {
+private fun PreviewPictureScreenContent_Light() {
     MakeColorTheme(isDarkTheme = false) {
         PreviewSurface {
-            RandomScreenContent(
-                viewModel = PreviewRandomViewModel(),
+            PictureScreenContent(
+                viewModel = PreviewPictureViewModel(),
                 onClickMenu = {},
                 onClickInfo = {},
                 onClickFloatingButton = {},
@@ -148,11 +165,11 @@ private fun PreviewRandomScreenContent_Light() {
 
 @Preview
 @Composable
-private fun PreviewRandomScreenContent_Dark() {
+private fun PreviewPictureScreenContent_Dark() {
     MakeColorTheme(isDarkTheme = true) {
         PreviewSurface {
-            RandomScreenContent(
-                viewModel = PreviewRandomViewModel(),
+            PictureScreenContent(
+                viewModel = PreviewPictureViewModel(),
                 onClickMenu = {},
                 onClickInfo = {},
                 onClickFloatingButton = {},
