+++ feature/makecolor/src/main/java/kosenda/makecolor/feature/makecolor/screen/SeekbarScreen.kt
@@ -1,11 +1,9 @@
-package kosenda.makecolor.view.screen
+package kosenda.makecolor.feature.makecolor.screen
 
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
@@ -20,65 +18,60 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import kosenda.makecolor.R
 import kosenda.makecolor.core.model.data.ColorData
-import kosenda.makecolor.core.ui.feature.common.ImagePicker
+import kosenda.makecolor.core.model.data.ColorType
 import kosenda.makecolor.core.ui.feature.common.button.FloatingAddButton
-import kosenda.makecolor.core.ui.feature.common.button.TonalButton
 import kosenda.makecolor.core.ui.feature.common.card.ColorValueTextsCard
 import kosenda.makecolor.core.ui.feature.common.card.DisplayColorCard
-import kosenda.makecolor.core.ui.feature.common.card.PaletteCircleCard
-import kosenda.makecolor.core.ui.feature.common.card.SelectImageCard
+import kosenda.makecolor.core.ui.feature.common.card.SeekbarsCard
+import kosenda.makecolor.core.ui.feature.common.card.SpinnerCard
 import kosenda.makecolor.core.ui.feature.common.topbar.TopBar
 import kosenda.makecolor.core.ui.feature.theme.MakeColorTheme
+import kosenda.makecolor.feature.makecolor.R
+import kosenda.makecolor.feature.makecolor.viewmodel.PreviewSeekbarViewModel
+import kosenda.makecolor.feature.makecolor.viewmodel.SeekbarViewModel
+import kosenda.makecolor.feature.makecolor.viewmodel.SeekbarViewModelImpl
 import kosenda.makecolor.feature.preview.PreviewSurface
-import kosenda.makecolor.view.content.GoogleAd
-import kosenda.makecolor.viewmodel.PictureViewModel
-import kosenda.makecolor.viewmodel.PictureViewModelImpl
-import kosenda.makecolor.viewmodel.PreviewPictureViewModel
 
 @Composable
-fun PictureScreen(
-    viewModel: PictureViewModelImpl,
+fun SeekbarScreen(
+    viewModel: SeekbarViewModelImpl,
     onClickMenu: () -> Unit,
     onClickInfo: () -> Unit,
     onClickFloatingButton: (ColorData) -> Unit,
     onClickDisplayColor: (ColorData) -> Unit,
+    googleAd: @Composable () -> Unit = {},
 ) {
-    PictureScreenContent(
+    SeekbarScreenContent(
         viewModel = viewModel,
         onClickMenu = onClickMenu,
         onClickInfo = onClickInfo,
         onClickFloatingButton = onClickFloatingButton,
         onClickDisplayColor = onClickDisplayColor,
+        googleAd = googleAd,
     )
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun PictureScreenContent(
-    viewModel: PictureViewModel,
+fun SeekbarScreenContent(
+    viewModel: SeekbarViewModel,
     onClickMenu: () -> Unit,
     onClickInfo: () -> Unit,
     onClickFloatingButton: (ColorData) -> Unit,
     onClickDisplayColor: (ColorData) -> Unit,
-    googleAd: @Composable () -> Unit = { GoogleAd() },
+    googleAd: @Composable () -> Unit = {},
 ) {
     val uiState by viewModel.uiState.collectAsState()
-    val context = LocalContext.current
-    val scrollState = rememberScrollState()
     val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior(rememberTopAppBarState())
-    val horizontalPadding = 16.dp
-    val launcher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.GetContent(),
-        onResult = {
-            it?.let { uri -> viewModel.makeBitmapAndPalette(uri = uri, context = context) }
-        },
+    val scrollState = rememberScrollState()
+    val colorTypeArray: Array<String> = arrayOf(
+        ColorType.RGB.name,
+        ColorType.CMYK.name,
+        ColorType.HSV.name,
     )
 
     Scaffold(
@@ -103,7 +96,7 @@ fun PictureScreenContent(
         Column(
             modifier = Modifier
                 .padding(it)
-                .padding(horizontal = horizontalPadding)
+                .padding(horizontal = 16.dp)
                 .nestedScroll(scrollBehavior.nestedScrollConnection)
                 .verticalScroll(scrollState),
         ) {
@@ -111,48 +104,37 @@ fun PictureScreenContent(
                 colorData = uiState.colorData,
                 onClickDisplayColor = onClickDisplayColor,
             )
-            if (uiState.bitmap == null) {
-                SelectImageCard(onClick = { launcher.launch("image/*") })
-            } else {
-                PaletteCircleCard(
-                    paletteColors = uiState.paletteColors,
-                    horizontalPadding = horizontalPadding,
-                    updateColorData = viewModel::updateColorData,
-                )
-                uiState.bitmap?.let { btm ->
-                    ImagePicker(
-                        btm = btm,
-                        horizontalPadding = horizontalPadding,
-                        updateColorData = viewModel::updateColorData,
-                    )
-                    Row(
-                        modifier = Modifier.padding(top = 16.dp),
-                    ) {
-                        TonalButton(
-                            text = stringResource(id = R.string.reset),
-                            painter = painterResource(id = R.drawable.baseline_cached_24),
-                            onClick = viewModel::resetImage,
-                        )
-                        Spacer(modifier = Modifier.weight(1f))
-                    }
-                }
-            }
-            ColorValueTextsCard(
+            SpinnerCard(
                 modifier = Modifier.padding(top = 8.dp),
+                selectedText = uiState.selectColorType.name,
+                categoryName = stringResource(id = R.string.category),
+                onSelectedChange = viewModel::updateSelectColorType,
+                displayItemList = colorTypeArray,
+            )
+            SeekbarsCard(
+                selectColorType = uiState.selectColorType,
+                onRGBColorChange = viewModel::updateColorData,
+                onCMYKColorChange = viewModel::updateColorData,
+                onHSVColorChange = viewModel::updateColorData,
                 colorData = uiState.colorData,
             )
-            Spacer(modifier = Modifier.height(150.dp))
+            ColorValueTextsCard(colorData = uiState.colorData)
+            Spacer(
+                modifier = Modifier
+                    .fillMaxWidth(1f)
+                    .height(150.dp),
+            )
         }
     }
 }
 
 @Preview
 @Composable
-private fun PreviewPictureScreenContent_Light() {
+private fun PreviewSeekbarScreenContent_Light() {
     MakeColorTheme(isDarkTheme = false) {
         PreviewSurface {
-            PictureScreenContent(
-                viewModel = PreviewPictureViewModel(),
+            SeekbarScreenContent(
+                viewModel = PreviewSeekbarViewModel(),
                 onClickMenu = {},
                 onClickInfo = {},
                 onClickFloatingButton = {},
@@ -164,11 +146,11 @@ private fun PreviewPictureScreenContent_Light() {
 
 @Preview
 @Composable
-private fun PreviewPictureScreenContent_Dark() {
+private fun PreviewSeekbarScreenContent_Dark() {
     MakeColorTheme(isDarkTheme = true) {
         PreviewSurface {
-            PictureScreenContent(
-                viewModel = PreviewPictureViewModel(),
+            SeekbarScreenContent(
+                viewModel = PreviewSeekbarViewModel(),
                 onClickMenu = {},
                 onClickInfo = {},
                 onClickFloatingButton = {},
